Return depends on language (#19)
* added possibly get data by language in categories

* added possibly get data by language in products

* sorted months

* fixed images

+++ backend/src/product/product.service.ts
@@ -2,13 +2,19 @@ import { Repository } from 'typeorm';
 import { InjectRepository } from '@nestjs/typeorm';
 import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 
+import { LanguageType } from '../decorators';
 import { ProductDto } from './dto/product.dto';
 import { ProductEntity } from './product.entity';
 import { FileService } from '../file/file.service';
 import { ProductServiceDto } from './dto/product.service.dto';
 import { CategoryService } from '../category/category.service';
 import { DuplicateService } from '../duplicate/duplicate.service';
-import { UpdateProductServiceDto } from './dto/update..product.service.dto';
+import { UpdateProductServiceDto } from './dto/update.product.service.dto';
+import {
+  sortedMonths,
+  getArrayByLanguage,
+  getObjectByLanguage,
+} from '../utils';
 
 @Injectable()
 export class ProductService {
@@ -19,6 +25,37 @@ export class ProductService {
     private duplicateService: DuplicateService,
     private fileService: FileService,
   ) {}
+
+  async findAll(
+    categoryId: string,
+    lang: LanguageType,
+  ): Promise<ProductEntity[]> {
+    const products = await this.productRepository.find({
+      where: { category_id: { id: categoryId } },
+    });
+
+    if (products?.length === 0) {
+      throw new HttpException('Products not found', HttpStatus.BAD_REQUEST);
+    }
+
+    return getArrayByLanguage(products, lang);
+  }
+
+  async findOne(id: string, lang?: LanguageType): Promise<ProductEntity> {
+    const product = await this.productRepository.findOneBy({
+      id,
+    });
+
+    if (!product) {
+      throw new HttpException('Product not found', HttpStatus.BAD_REQUEST);
+    }
+
+    if (lang) {
+      return getObjectByLanguage(product, lang);
+    }
+
+    return product;
+  }
 
   async create(
     productDto: ProductServiceDto | ProductDto,
@@ -33,7 +70,7 @@ export class ProductService {
     }
 
     await this.duplicateService.check(this.productRepository, {
-      name: productDto.name,
+      name_en: productDto.name_en,
       name_ua: productDto.name_ua,
     });
 
@@ -42,36 +79,16 @@ export class ProductService {
       product.img_url = imgUrl as string;
     }
 
+    if (Array.isArray(productDto.months)) {
+      productDto.months = sortedMonths(productDto.months);
+    }
+
     const newProduct = await this.productRepository.create(
       product as ProductServiceDto,
     );
     newProduct.category_id = category;
 
     return this.productRepository.save(newProduct);
-  }
-
-  async findAll(categoryId: string): Promise<ProductEntity[]> {
-    const products = await this.productRepository.find({
-      where: { category_id: { id: categoryId } },
-    });
-
-    if (products?.length === 0) {
-      throw new HttpException('Products not found', HttpStatus.BAD_REQUEST);
-    }
-
-    return products;
-  }
-
-  async findOne(id: string): Promise<ProductEntity> {
-    const product = await this.productRepository.findOneBy({
-      id,
-    });
-
-    if (!product) {
-      throw new HttpException('Product not found', HttpStatus.BAD_REQUEST);
-    }
-
-    return product;
   }
 
   async update(
@@ -83,8 +100,8 @@ export class ProductService {
 
     const fieldsToCheck = {};
 
-    if (product.name !== productDto.name) {
-      fieldsToCheck['name'] = productDto.name;
+    if (product.name_en !== productDto.name_en) {
+      fieldsToCheck['name_en'] = productDto.name_en;
     }
 
     if (product.name_ua !== productDto.name_ua) {
@@ -99,12 +116,16 @@ export class ProductService {
       );
     }
 
+    if (Array.isArray(productDto.months)) {
+      productDto.months = sortedMonths(productDto.months);
+    }
+
     if (file) {
       const imgUrl = await this.fileService.create(file);
       productDto.img_url = imgUrl;
     }
 
-    if (!file) {
+    if (!file && product.img_url) {
       await this.fileService.delete(product.img_url);
     }
